친구 서비스 완료, 계정삭제여부 수정완료

+++ controllers/ProfileController.php
@@ -15,9 +15,46 @@ try {
          * 마지막 수정 날짜 : 20.08.31
          */
         case "getProfile":
-
             http_response_code(200);
-            $res->result->profile = userProfile($vars["userIdx"]);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            $data=getDataByJWToken($jwt,JWT_SECRET_KEY);
+            if (!isValidHeader($jwt, JWT_SECRET_KEY) || !isJwtSaved($jwt,1)) {
+                $res->isSuccess = FALSE;
+                $res->code = 201;
+                $res->message = "유효하지 않은 토큰입니다";
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                addErrorLogs($errorLogs, $res, $req);
+                break;
+            }
+
+            if(!isValidUserid($vars["userIdx"])){
+                $res->isSuccess = FALSE;
+                $res->code = 210;
+                $res->message = "없는 계정입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
+            if(getUserStatus($data->userId)==2) {
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
+            if(getUserStatus($vars["userIdx"])==2){
+                $res->isSuccess = FALSE;
+                $res->code = 200;
+                $res->message = "이 프로필은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
+            $res->result->profile = userProfile($vars["userIdx"], $data->userId);
             $res->result->friendsCount = userProfileFriendsCount($vars["userIdx"]);
             $res->result->friends = userProfileFriends($vars["userIdx"]);
             $res->isSuccess = TRUE;
@@ -43,8 +80,18 @@ try {
                 addErrorLogs($errorLogs, $res, $req);
                 break;
             }
+
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
             http_response_code(200);
-            $res->result = openModifyPage($data->userId); // 수정필요
+            $res->result = openModifyPage($data->userId);
             $res->isSuccess = TRUE;
             $res->code = 100;
             $res->message = "상세 프로필 조회 성공";
@@ -68,6 +115,16 @@ try {
                 echo json_encode($res, JSON_NUMERIC_CHECK);
                 break;
             }
+
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
             if(is_null($req->contents)||$req->contents==""){
                 $res->isSuccess = FALSE;
                 $res->code = 200;
@@ -78,7 +135,7 @@ try {
             }
             if(mb_strlen($req->contents, "UTF-8")>100){
                 $res->isSuccess = FALSE;
-                $res->code = 201;
+                $res->code = 202;
                 $res->message = "100자 이내로 입력해주세요.";
 
                 echo json_encode($res, JSON_NUMERIC_CHECK);
@@ -111,6 +168,15 @@ try {
             }
             http_response_code(200);
 
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
 
             deleteIntroduce($data->userId); // 수정필요
             $res->isSuccess = TRUE;
@@ -136,6 +202,15 @@ try {
                 break;
             }
             http_response_code(200);
+
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
 
             modifyHobby($req->contents,$data->userId); // 수정필요
             $res->isSuccess = TRUE;
@@ -164,9 +239,27 @@ try {
             }
             $offset = $_GET['offset']*50;
 
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
+            if(getUserStatus($vars["userIdx"])==2){
+                $res->isSuccess = FALSE;
+                $res->code = 200;
+                $res->message = "이 프로필은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
             if(userProfileFriendsCount($vars['userIdx'])['friendsCount']<$offset){
                 $res->isSuccess = TRUE;
-                $res->code = 200;
+                $res->code = 210;
                 $res->message = "더 이상 친구가 없습니다.";
 
                 echo json_encode($res, JSON_NUMERIC_CHECK);
@@ -174,7 +267,7 @@ try {
 
             }
 
-            $res->result = getAllFriends($vars['userIdx'], $offset); // 수정필요
+            $res->result = getAllFriends($vars['userIdx'], $offset);
             $res->isSuccess = TRUE;
             $res->code = 100;
             $res->message = "친구목록 조회 성공";
@@ -182,10 +275,45 @@ try {
             echo json_encode($res, JSON_NUMERIC_CHECK);
             break;
 
+        /*
+        * API No. 6
+        * API Name : 친구 검색
+        * 마지막 수정 날짜 : 20.09.01
+        */
+        case "getSearchedFriend":
+            http_response_code(200);
+            $jwt = $_SERVER["HTTP_X_ACCESS_TOKEN"];
+            $data=getDataByJWToken($jwt,JWT_SECRET_KEY);
+            if (!isValidHeader($jwt, JWT_SECRET_KEY) || !isJwtSaved($jwt,1)) {
+                $res->isSuccess = FALSE;
+                $res->code = 201;
+                $res->message = "유효하지 않은 토큰입니다";
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
+
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
+            $res->result = getSearchedFriend($vars['userIdx'], $_GET['search']);
+            $res->isSuccess = TRUE;
+            $res->code = 100;
+            $res->message = "친구 조회 성공";
+
+            echo json_encode($res, JSON_NUMERIC_CHECK);
+            break;
+
 
         /*
-        * API No. 5
-        * API Name : 친구목록(페이징)
+        * API No. 7
+        * API Name : 디테일페이지 API
         * 마지막 수정 날짜 : 20.09.01
         */
         case "getMyDetailPage":
@@ -200,12 +328,21 @@ try {
                 break;
             }
 
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
             $res->result->work = getMyDetailWork($data->userId); // 수정필요
             $res->result->School = getMyDetailSchool($data->userId); // 수정필요
-            $res->result->living = getMyDetail($data->userId); // 수정필요
+            $res->result->info = getMyDetail($data->userId); // 수정필요
             $res->isSuccess = TRUE;
             $res->code = 100;
-            $res->message = "친구목록 조회 성공";
+            $res->message = "상세정보 페이지 조회 성공";
 
             echo json_encode($res, JSON_NUMERIC_CHECK);
             break;
@@ -230,7 +367,7 @@ try {
             move_uploaded_file($_FILES['uploaded_file']['tmp_name'], $file_path);
             if (isValidFileName($file_calling)) {
                 $res->isSuccess = FALSE;
-                $res->code = 205;
+                $res->code = 206;
                 $res->message = "파일이름이 중복입니다.";
                 echo json_encode($res, JSON_NUMERIC_CHECK);
                 break;
@@ -257,6 +394,15 @@ try {
                 break;
             }
 
+            if(getUserStatus($data->userId)==2){
+                $res->isSuccess = FALSE;
+                $res->code = 205;
+                $res->message = "로그인 한 계정은 비활성된된 유저입니다.";
+
+                echo json_encode($res, JSON_NUMERIC_CHECK);
+                break;
+            }
+
             $name = $_POST['name'];
 
             $file_path = './photos/' . $name . '.jpg';
@@ -266,7 +412,7 @@ try {
 
             if (isValidFileName($file_calling)) {
                 $res->isSuccess = FALSE;
-                $res->code = 205;
+                $res->code = 206;
                 $res->message = "파일이름이 중복입니다.";
                 echo json_encode($res, JSON_NUMERIC_CHECK);
                 break;
